+++ scripts/build_renewable_profiles.py
@@ -245,7 +245,7 @@ if __name__ == "__main__":
     countries = snakemake.config["countries"]
     paths = snakemake.input
     nprocesses = snakemake.config["atlite"].get("nprocesses")
-    noprogress = not snakemake.config["atlite"].get("show_progress", True)
+    noprogress = not snakemake.config["atlite"].get("show_progress", False)
     config = snakemake.config["renewable"][snakemake.wildcards.technology]
     resource = config["resource"]
     correction_factor = config.get("correction_factor", 1.0)